Fixed a warning message about getTrack() which the compiler thinks is unsigned and it really is an integer
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10914 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/document/io/LilyPondExporter.cpp
@@ -746,8 +746,11 @@ LilyPondExporter::write()
     }
     timeT compositionStartTime = (*i)->getStartTime();
     timeT compositionEndTime = (*i)->getEndMarkerTime();
+    // there is something amiss with the compiler since in CompositionTimeSliceAdapter
+    // getTrack() returns an integer and not unsigned. A minus value is used for an
+    // illegal flag, so the test has meaning. I will simply cast it. Ilan.
     for (; i != m_composition->end(); ++i) {
-        if (compositionStartTime > (*i)->getStartTime() && (*i)->getTrack() >= 0) {
+        if (compositionStartTime > (*i)->getStartTime() && (int)(*i)->getTrack() >= 0) {
             compositionStartTime = (*i)->getStartTime();
         }
         if (compositionEndTime < (*i)->getEndMarkerTime()) {
